Fixat konto 5021

+++ DataLayer/DBKlasser/PersonalProdukt.cs
@@ -18,7 +18,7 @@ namespace DataLayer
         [Column(Order = 1)]
         public string Produkt_ProduktID { get; set; }
 
-        public int Placeringsandel { get; set; }
+        public decimal Placeringsandel { get; set; }
 
         public virtual Personal Personal { get; set; }
     }
